requirements.txt
alembic==0.9.9
blinker==1.4
click==6.7
dominate==2.3.1
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SimpleMDE==0.3.0
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
markdown2==2.3.5
MarkupSafe==1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
PyJWT==1.6.1
python-dateutil==2.7.2
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.2.6
uuid==1.30
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.1
